BS-692_ingesting_with_estimate_id_null_db_changes2

+++ Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/Entities/SpotExceptions/Unposted/OutOfSpecUnpostedNoReelRosterDto.cs
@@ -31,7 +31,7 @@
         /// Gets or sets the estimate identifier.
         /// </summary>
         /// <value>The estimate identifier.</value>
-        public long? EstimateId { get; set; }
+        public int? EstimateId { get; set; }
 
         /// <summary>
         /// Gets or sets the ingested by.
