+++ Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/Entities/SpotExceptions/Unposted/SpotExceptionsUnpostedNoReelRosterDto.cs
@@ -6,7 +6,7 @@
         public string HouseIsci { get; set; }
         public int Count { get; set; }
         public System.DateTime ProgramAirTime { get; set; }
-        public long? EstimateId { get; set; }
+        public int? EstimateId { get; set; }
         public string IngestedBy { get; set; }
         public System.DateTime IngestedAt { get; set; }
         public int IngestedMediaWeekId { get; set; }
